+++ pkg/gtm/common.go
@@ -6,7 +6,7 @@ import (
 )
 
 // default schema version
-var schemaVersion = "1.4"
+var schemaVersion = "1.6"
 
 // internal method to set version. passed in as string
 func setVersionHeader(req *http.Request, version string) {